refactor: lifting popButtonCss
