+++ packages/menu/src/style.tsx
@@ -1,7 +1,22 @@
 export * from "./styles"
 import { css } from "@emotion/react"
 import { globalColor, illaPrefix } from "@illa-design/theme"
+import { SerializedStyles } from "@emotion/serialize"
+import { Theme } from "./interface"
 
-export function applyMenu() {
-
+export function applyPopButtonCss(theme: Theme = "light"): SerializedStyles {
+  return css`
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    box-shadow: 0 4px 10px ${globalColor(`--${illaPrefix}-gray-09`)};
+    padding: 0 12px;
+    box-sizing: border-box;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 16px;
+    background: ${theme === "dark"
+      ? globalColor(`--${illaPrefix}-gray-02`)
+      : "none"};
+  `
 }
