+++ packages/menu/src/styles/item-group.tsx
@@ -2,6 +2,7 @@ import { css } from "@emotion/react"
 import { SerializedStyles } from "@emotion/serialize"
 import { globalColor, illaPrefix } from "@illa-design/theme"
 import { Theme } from "../interface"
+import { applyPopButtonCss } from "../style"
 
 export function applyItemGroupCss(theme: Theme = "light"): SerializedStyles {
   const color =
@@ -19,20 +20,6 @@ export function applyGroupTitleCss(
   isCollapse?: boolean,
   theme: Theme = "light"
 ): SerializedStyles {
-  const popButtonCss = css`
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    box-shadow: 0 4px 10px ${globalColor(`--${illaPrefix}-gray-09`)};
-    padding: 0 13px;
-    box-sizing: border-box;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 16px;
-    background: ${theme === "dark"
-      ? globalColor(`--${illaPrefix}-gray-02`)
-      : "none"};
-  `
 
   const padding = isCollapse ? 16 : 24
 
@@ -50,6 +37,6 @@ export function applyGroupTitleCss(
   return css([
     itemGroupTitleCss,
     !isHorizontal && overflowCss,
-    isPopButton && popButtonCss,
+    isPopButton && applyPopButtonCss(theme),
   ])
 }
