+++ packages/menu/src/styles/item.tsx
@@ -2,6 +2,7 @@ import { css } from "@emotion/react"
 import { SerializedStyles } from "@emotion/serialize"
 import { globalColor, illaPrefix } from "@illa-design/theme"
 import { Theme } from "../interface"
+import { applyPopButtonCss } from "../style"
 
 export function applyItemTitleCss(): SerializedStyles {
   const titleEllipsis = css`
@@ -50,8 +51,8 @@ export function applyItemCss(
 
   const hoverCss = css`
     &:hover {
-      cursor: pointer;
-      background-color: ${themeColor[theme].hoverBg};
+    cursor: pointer;
+    background-color: ${themeColor[theme].hoverBg};
     }
   `
 
@@ -65,30 +66,17 @@ export function applyItemCss(
     background-color: ${themeColor[theme].selectedBg};
   `
 
-  const popButtonCss = css`
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    box-shadow: 0 4px 10px ${globalColor(`--${illaPrefix}-gray-09`)};
-    padding: 0 14px;
-    box-sizing: border-box;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 16px;
-    background: ${theme === "dark" ? globalColor(`--${illaPrefix}-gray-02`) : "none"}
-  `
-
   const horizontalSelectedCss = css`
     background: none;
     &:after {
-      content: "";
-      display: block;
-      height: 2px;
-      width: 100 %;
-      bottom: 0;
-      left: 0;
-      background-color: ${themeColor[theme].horizontalSelectedBg};
-      position: absolute;
+    content: "";
+    display: block;
+    height: 2px;
+    width: 100 %;
+    bottom: 0;
+    left: 0;
+    background-color: ${themeColor[theme].horizontalSelectedBg};
+    position: absolute;
     }
   `
 
@@ -104,6 +92,6 @@ export function applyItemCss(
     ${isDisabled ? disabledCss : hoverCss};
     ${isSelected && selectedCss};
     ${isSelected && isHorizontal && horizontalSelectedCss};
-    ${isPopButton && popButtonCss};
+    ${isPopButton && applyPopButtonCss(theme)};
   `
 }
