+++ packages/menu/src/sub-menu/pop.tsx
@@ -83,7 +83,7 @@ export const Pop = forwardRef<HTMLDivElement, SubMenuProps>((props, ref) => {
   const mergedTriggerProps = {
     colorScheme: theme === "light" ? "white" : "gray",
     position: (needPopOnBottom ? "bl" : "rt") as TriggerProps["position"],
-    showArrow: variant !== "pop",
+    showArrow: mode === "popButton" || variant !== "pop",
     withoutPadding: true,
     clickOutsideToClose: true,
     ...triggerProps,
